# Separate size and capacity of LGDO ArrayLike objects