--- a/docs/source/notebooks/LH5Files.ipynb
+++ b/docs/source/notebooks/LH5Files.ipynb
@@ -189,8 +189,8 @@
    "source": [
     "from lgdo.lh5 import LH5Iterator\n",
     "\n",
-    "for lh5_obj, entry, n_rows in LH5Iterator(lh5_file, \"geds/raw/energy\", buffer_len=20):\n",
-    "    print(f\"entry {entry}, energy = {lh5_obj} ({n_rows} rows)\")"
+    "for lh5_obj, entry in LH5Iterator(lh5_file, \"geds/raw/energy\", buffer_len=20):\n",
+    "    print(f\"entry {entry}, energy = {lh5_obj} ({len(lh5_obj)} rows)\")"
    ]
   },
   {
@@ -211,7 +211,7 @@
     "from lgdo.lh5 import LH5Store\n",
     "\n",
     "store = LH5Store(keep_open=True)  # with keep_open=True, files are kept open inside the store\n",
-    "store.read(\"geds/raw\", lh5_file)  # returns a tuple: (obj, n_rows_read)"
+    "store.read(\"geds/raw\", lh5_file)\n"
    ]
   },
   {