--- a/docs/source/notebooks/LH5Files.ipynb
+++ b/docs/source/notebooks/LH5Files.ipynb
@@ -212,8 +212,10 @@
    "source": [
     "from lgdo.lh5 import LH5Store\n",
     "\n",
-    "store = LH5Store(keep_open=True)  # with keep_open=True, files are kept open inside the store\n",
-    "store.read(\"geds/raw\", lh5_file)\n"
+    "store = LH5Store(\n",
+    "    keep_open=True\n",
+    ")  # with keep_open=True, files are kept open inside the store\n",
+    "store.read(\"geds/raw\", lh5_file)"
    ]
   },
   {