--- a/docs/source/notebooks/LH5Files.ipynb
+++ b/docs/source/notebooks/LH5Files.ipynb
@@ -191,8 +191,8 @@
    "source": [
     "from lgdo.lh5 import LH5Iterator\n",
     "\n",
-    "for lh5_obj, entry in LH5Iterator(lh5_file, \"geds/raw/energy\", buffer_len=20):\n",
-    "    print(f\"entry {entry}, energy = {lh5_obj} ({len(lh5_obj)} rows)\")"
+    "for lh5_obj in LH5Iterator(lh5_file, \"geds/raw/energy\", buffer_len=20):\n",
+    "    print(f\"energy = {lh5_obj} ({len(lh5_obj)} rows)\")"
    ]
   },
   {