--- a/src/lgdo/types/array.py
+++ b/src/lgdo/types/array.py
@@ -146,6 +146,10 @@ class Array(LGDO):
 
     def insert(self, i: int, value: int | float) -> None:
         "Insert value into row i (with copy)"
+        if i > len(self):
+            msg = f"index {i} is out of bounds for array with size {len(self)}"
+            raise IndexError(msg)
+        
         value = np.array(value)
         if value.shape == self.shape[1:]:
             self.resize(len(self) + 1)