… src/lgdo/types/lgdo.py
… src/lgdo/types/table.py
… src/lgdo/types/vectorofvectors.py
… src/lgdo/types/vovutils.py
… tests/compression/conftest.py
… tests/compression/test_radware_sigcompress.py
… tests/lh5/conftest.py
… tests/lh5/test_core.py
… tests/lh5/test_lh5_iterator.py
… tests/lh5/test_lh5_store.py
… tests/lh5/test_lh5_write.py
… tests/test_cli.py
… tests/types/test_array.py
… tests/types/test_histogram.py
… tests/types/test_table.py
… tests/types/test_vectorofvectors.py