@ggmarshall 2025-03-16
Ok I have everything working now I believe on the pygama side and have run the data production with this, happy for this to be merged